Fix problem with cargo trying to download the registry while --frozen.
Cargo needs a `.cargo/config` to tell it that we're only using offine (vendored)
dependencies.  It will look for that directory as a subdirectory of the current
working directory, or of any directory upwards from the current working
directory.  Copying it into build-tree-tor didn't fix the issue for some reason,
but copying it into $HOME does.

[isis grimaces while waiting for weasel's distaste at copying something to $HOME]

 * FIXES part of #22768: https://bugs.torproject.org/22768